+++ examples/blackbox_jaspr_example/lib/account_auth_profile_async/ui/services_page.dart
@@ -116,7 +116,7 @@ class _GlobalState extends State<_Global> {
           _selectedService,
           input: (d) => d.whenReady(_servicesLoader),
         )
-        .build(start: true);
+        .build(start: true, trace: true);
   }
 
   @override
@@ -236,7 +236,7 @@ class _ServiceState extends State<_Service> {
           _profile!,
           input: (d) => d.whenReady(_auth!),
         )
-        .build(start: true);
+        .build(start: true, trace: true);
   }
 
   @override