+++ examples/blackbox_jaspr_example/lib/counter_sync_async/boxes/step_config.dart
@@ -6,7 +6,7 @@ class StepConfig extends NoInputBox<int> {
   StepConfig();
 
   @override
-  int computeValue(int? previous) => _step;
+  int compute(int? previous) => _step;
 
   void inc() => action(() => _step++);